Added support for uvisor_box_context
`uvisor_box_context` will end up in SRAM (`.bss`) at a position that is not
known at (uVisor) compile time.

So, the address of `uvisor_box_context` will be read at runtime from a const
symbol in Flash, set by the UVISOR_SET_MODE macro.

This is basically the same approach that was used for `__uvisor_mode`.

In this commit:
- added symbol to `.s` file for `uvisor-lib` that holds the address of the
  context sp
- added field in `__uvisor_config` struct where uvisor fetches linker script
  symbols
- added code to read `uvisor_box_context` at runtime and to switch between
  different boxes context stack pointers during a secure gateway