Added function to recover from memory accesses faults
In the very particular case in which a memory access is attempted using ldrX or
strX instructions using r0 as address and r1 as value (str only), we allow the
memory access to go through (if an ACL allows it, TODO).

MemManage faults (ARM MPU) or BusFaults (Freescale MPU) can than return to the
instruction following the faulting one.