Disabling bus write buffering for K64F
This is a temporary fix.

We want to be able to recover from MPU memory exceptions, that in the K64F
result in bus faults. Imprecise bus faults are not trivial to recover, because
they can be imprecise. This temporary fix disable write buffering so bus faults
are always precise, despite the system being a bit slower.