Added imprecise bus fault recovery (all platforms)
With imprecise faults both the stacked pc (sp[6]) and the syndrome register BFAR
are not valid. Being able to recover from an imprecise bus fault is important
especially with the K64F, where MPU faults are bus fauts.

The stacked pc is used as starting point for a backward search that tries to
find the offending instructions. If this is found, the memory acces is performed
and the exception returns.

A dsb and 3 nop instructions are added to make sure nothing is executed before
the fault occurs.

Note: this only applies for "exception" registers in the K64F, for which nothing
can be done to enable unprivileged access (no MPU region, no PACRx register nor
any specific flag can be set to enable it). In those cases read and write
operations have been transformed into custom ldrX/strX operations, forged for
this recovery to be possible.