uvisor-lib release now generates an object file
Currently an input .S file is pre-processed into a gcc-style asm file (.s),
which is not compatible with armcc. To avoid code duplication, the release
process has been made easier by pre-processing and compiling uvisor into
uvisor-lib all in one place (here) and then only publishing the object file in
the yotta module.

A side-effect advantage is that binary blobas will be hardware- but not
toolchain-specific.