Fixed bug in sanity check
New linker script symbol __uvisor_bss_boxes_end represents the end address
for the memory reserved to secure boxes' private stacks+contexts.

The previous check used __uvisor_bss_start, which was in fact the same exact symbol,
but with a misleading name: it was assumed that "the start of the uVisor-protected bss"
was always be right after the space reserved to secure boxes stacks.