UVISOR_PRESENT is now YOTTA_CFG_UVISOR_PRESENT
The symbol change implies that it is not necessary anymore to change platforms.h to add support to a platform; instead,
the target that supports uVisor will define that in its configuration.

Remember that "UVISOR_PRESENT" means that uVisor is supported and the linker script, startup routines, etc. have been
adapted to work with uVisor (see porting guide). When UVISOR_PRESENT is 1, uVisor can still be disabled (in fact, it is
currently disabled by default).