Fix bug on 0-args secure gateway
A secure gateway with 0 arguments gave a compilation error since macro
overloading was not done properly.

GNU CPP mandates that in order to have comma swallowing when using
```C
, ## __VA_ARGS__
```
__VA_ARGS__ must be left out completely; if it's not left out the extension
does not apply, even if __VA_ARGS__ is an empty token.

The bug was similar to the following example:
```C
\#define MACRO1(a0, ...) MACRO2(a0, __VA_ARGS__)                    // bug here
\#define MACRO2(a0, ...) FINALLY_DO_SOMETHING(a0, ##__VA_ARGS__)
```
Not using the GNU CPP extension in MACRO1 makes the one in MACRO2 ineffective,
since an empty token (not a missing token) is sent to MACRO2.

In uVisor the macro `UVISOR_SVC` corresponds to `MACRO1` in the example above; it
was passing `__VA_ARGS__` without the correct GNU CPP extensions, hence triggering
and empty token in the underlying macro (`UVISOR_MACRO_REGS_ARGS`).