Added check for max priority in vIRQ APIs
Since we offset all interrupt priorities by `UNVIC_MIN_PRIORITY` we
want to make sure that the priority field does not overflow.

Added:
- sanity check on available number of interrupts priority levels
- default setting for priority grouping (this is set by default
  after reset but it is needed anyway as it might happen that
  `SystemInit()` changed it before `uvisor_init()`
- runtime check for `vIRQ_SetPriority()`

Note: the same runtime check is not needed for `vIRQ_SetVector()`
as the minimum priority is set by default and the `unvic_init()`
routine already checked that this is legit.