vIRQ_SetVector turned into an actual function
This allows the function to exist as an external symbol and to be
used accordingly (for example, in a secure gateway)

The reason why this is needed is that although `vIRQ_SetVectorX` has a wider
scope than `vIRQ_SetVector`, the latter will be preferred for backward
compatibility reasons for the foreseeable future (it has the same signature
of `NVIC_SetVector`).

The change introduced here allows the API to share the same properties
of the similar APIs in the vIRQ module: an explicit symbol, direct
secure-gateway-ability.