Disable interrupt when releasing ownership of IRQn
If an IRQn is released (de-registered) but left enabled, a spurious interrupt
could be served.

When serving an interrupt we trust the information that we hold for the IRQn, that is,
the ID of the interrupt owner and the ISR vector. If an interrupt is released but
it is still enabled, when it gets served the default handler (__unvic_svc_cx_in) will
see the following:
ID of owner   : 0
associated ISR: 0x0
and it will serve it without any further check (some default checks are implemented
in unvic_default_check() but they do not cover this corner case).

See comment in https://github.com/ARMmbed/uvisor/blob/master/core/system/src/unvic.c#L322