Export symbol of max priority level allowed (users)
With the classic NVIC module, a user would typically use `__NVIC_PRIO_BITS` to
determine the max priority level available. This commit makes sure that a
similar macro is exported for vIRQ users. Currently the symbol still relies on
`__NVIC_PRIO_BITS`, but this could change in the future if more complex
virtualisation is implemented.

Note that the minimum NVIC priority must be exported as well, but the name
has been prefixed with a double underscore to discourage users from using it
explicitly (it would make no sense as the NVIC module cannot be accessed
directly).