Update build system after refactoring
The changes introduced here to the main Makefile allow a more
centralised experience when porting a device family to uVisor.

* The symbols for CPU/ARCH have been removed. The build system is now
* "project"-
  "configuration"-oriented, where: ** A project is the folder where a
whole device family is ported. It is located in `platform/` ** A
configuration is the combination of several pre-processor macros
definitions that trigger the release of a uVisor binary. For each
project (ported family) there will be as many uVisor binaries as the
configurations. The uVisor refactoring allows the least possible number
of binaries to be generated, leveraging similarities between targets.
* The release script has been updated to generate one binary for each
  configuration. Whether all configurations are automatically generated
with one signle command is up to the platform port. For example, a
recursive Makefile can be used that iterates over all the possible
configurations.