Re-add support for K64F after refactoring
This port now complies with the new device-family-oriented folder
structure. Out of the Kinetis family, we currnetly only support the
K64F, although this is easily expandable.

The commit also changes the `ARCH_MPU_MK64F` definition to the more
general `ARCH_MPU_KINETIS`, since that MPU configuration is common to
the whole Kinetis family.