Add IRQ context switch fallback for disabled.cpp
The changes introduced here allow programs written using uVisor APIs and
secure box contexts (at the same time) to work even when uVisor is
present, but disabled.

A custom vector table is kept, which is used to register/de-register
interrupts. The actual interrupt handler in the mbed NVIC table is a
default handler managed by disabled.cpp that performs a context switch.

In this way box contexts are still switched even with an IRQ.