Add warning in README after refactoring.
The README now includes a warning on the debug features that have been
temporarily removed from uVisor after the refactoring. It also includes
instructions on how to fetch a fallback branch that is kept for users to
use the features in a previous uVisor version.

Note: The README on that branch (`uvisor_with_memory_map_debug`) does
not include this warning. It will present the old building instructions
normally as they apply to the old version of uVisor.