Make debug/release build rules generic
This commit moves the debug and release build rules for GNU make to the
central Makefile.rules file.

It is now not needed any more to provide custom rules to build platform
families. Instead, by specifying configurations in the form:
```make
CONFIGURATIONS:= \
    CONFIGURATION_${SOMETHING} \
    ... \
    CONFIGURATION_${SOMETHING_ELSE}
```
the main Makefile is able to build all binaries in one single nested
make.

Note: For this change, the clean rule has been reverted to its original
behaviour -- it deletes all the content of the `release/${family}/`
folder.

Fixes: 7d60c75 ("Re-add support for K64F after refactoring")
Fixes: e44d0c8 ("Re-add support for STM32F429 after refactoring")
Fixes: 15227c2 ("Do not remove the full release folder for a family")
Fixes: #152 ("Not possible to release multiple configurations for a platform")