Remove host-OS-specific files
`unsupported.c` Will be added by a host OS if there is a need to build
an application that links to uVisor on an unsupported platform. The file
will only need to implement an empty hook for `uvisor_init()`.
This file will be added by the host OS if support is needed for
unsupported platforms.

`unsupported.h` Will also include re-definitions of uVisor macros or
symbols to make sure that special mechanisms like SVCall redirections
are avoided on unsupported platforms.

`override.h` Will be added by a host OS if there is a need to override
NVIC APIs with vIRQ ones in source files that should not be changed. A
single-point change (`#include "override.h"`) overwrites the NVIC APIs.