Make UVISOR_PRESENT symbol non-yotta-specific
Some of the uVisor files were using YOTTA_CFG_UVISOR_PRESENT as a symbol
to select the correct version of a pre-processor macro (supported vs.
unsupported), but this symbol is only available when yotta is used to
compile uVisor.

This commit generalizes the symbol to UVISOR_PRESENT. A related change
in uvisor-lib must create the glue logic to convert YOTTA_CFG_* symbols
to uVisor-specific ones. This currently applies to:
    * UVISOR_PRESENT
    * ARCH_MPU_{ARMv7M, KINETIS}

The Makefile has been updated to define UVISOR_PRESENT by default (used
when compiling the release library).