export-table: Add export table
Whenever uVisor has a need to communicate symbols contained within the uVisor
binary to the application, it can use the export table.

uvisor_post_init() is also added. The system startup must call this to
initialize uvisor-lib things. In the initial implementation of
uvisor_post_init(), we check that the export table is what we expect from
uVisor. This keeps applications and uVisor binaries from getting out of sync
via a run-time check.