box-main: Implement uVisor box main feature
Make use of the recently added box init mechanism to implement the
uVisor box main feature. uVisor now deprivileges and calls the
uVisor-lib-provided box init function for each box. On CMSIS RTOS, the
uVisor-lib-provided box init function currently creates a box main
thread.

The box main thread stack is separate from the box stack. This is
because the thread must be created to use a different stack than the
stack `osCreateThread()` is called from, as context information is saved
to the thread stack by the call to `osCreateThread()`. The box main
thread stack is created from the process heap (which is memory private
to the box).