Remove wrong volatile cast from unsupported.h
The volatile cast was initially introduced for consistency with the
supported case. The supported case needed the volatile vast (volatile
pointer) because of a compiler issue (GCC), where optimizations changed
our inline assembly code. See #173 for more details, and #172 for the
fix (supported case).

The volatile was mistakenly introduced in unsupported.h. Fixes:

* 61d0a1c "Turn ADDRESS_WRITE into a static inline function"
* ad7271a "Add missing volatile cast"