vIRQ_{Dis,En}ableAll: Fix bug of never-enabled IRQs
When nesting multiple pairs of disable-all/enable-all function calls,
the last call to vIRQ_EnableAll() was not re-enabling all interrupts.

This was due to the fact that if multiple vIRQ_DisableAll() were called,
the wrong state `was_enabled = false` was set, even if the IRQ was
previously enabled.

Upon restoring the state, those IRQs were not re-enabled.

IRQs that were enabled during the critical section were correctly
enabled, because their state was unaffected.

This also improves performance because in the previous implementation
IRQs were always iterated over even if the counter was more than 0,
which is useless.

Fixes:

* bcf5fbe "Disable all IRQs: Added virtualised APIs".