Register gateway: New macros for API translation
The new macros update the way the secure gateway API turns into a binary
structure in flash.

In particular, this commit fixes a limitation of the previous API for
which only explicit hex addresses could be used in the `addr` argument.

Now any address is supported.

New APIs are also provided that allow common operations on bits to be
performed via the uVisor register gateway. All those APIs are turned
into a `uvisor_{read,write}` function call with the appropriate
parameters.

Fixes:

* #90 "Add symbol support to uvisor_read & uvisor_write APIs".