Make access to MPU regions 0-2 more static
The lower 3 MPU regions are used for the background region and allowing
various accesses for uVisor and user code.
These 3 regions are initialized once and then never switched and should
therefore not take part in the vmpu_mem_switch code to make it easier
to reason about the usage of the remaining 8 MPU regions.

Note that the MPU region to unlock uVisor SRAM is unnecessary, since it
has the same priviledges as the background region and has been removed.