K64F vMPU: Use the MPU fault address
When recovering from an MPU fault on K64F, we were using the bus fault
syndrome register. This register does not contain a valid address when
the bus fault is imprecise.

On the other hand, the syndrome register for the MPU module always
contains a valid register when an MPU fault occur.

This commit introduces 2 `inline` APIs to fetch and clear the fault
status for a given slave port, and uses them to recover from a potential
MPU fault.

This fixes a bug found in release builds of our `blinky` example, where
some faults were imprecise.

Fixes:
* 2c565f8 Implement recovering from page heap faults