ARMv7-M MPU: Fix missing box 0 ACLs
When searching for ACLs upon a MemManage fault on ARMv7-M MPUs, the
existing code was returning too early when an ACL could not be found for
the currently active box. In this way, shared ACLs from box 0 were never
used.

This commit fixes the code by returning early only when the region is
**found** for the current box, otherwise trying again for box 0.
Eventually, if no ACL is found (i.e., no `if` is entered), the function
returns `NULL`.