Debug box: Do not reboot by default
The `debug_halt_error` handler should not implicitly infer a reboot of
the system.

The debug box handler can reboot, if it thinks it's necessary. The newly
introduced vIRQ_SystemReset API can be used for this purpose.

By default, the debug handler will trigger a call to `uvisor_error`, a
user-level API to trigger a halt. The halt will have the newly
introduced reason `DEBUG_BOX_HALT`.