Save some flash space by consolidating halts
This helps reduce the flash size of our debug build at the expense of
the suitability of the debug build for debugging.

An issue at https://github.com/ARMmbed/uvisor/issues/337 tracks our
progress towards solving this problem in a way that does not reduce the
suitability of the debug build for debugging.