Place all of uVisor into secure alias
This is required to boot uVisor on the secure side but still have its
binary readable on the non-secure side.

When uVisor boots the hardware is in secure mode with the SAU switched
off, allowing secure execution from anywhere. When switching the SAU on,
if we are executing from non-secure memory, a secure fault occurs,
therefore we must execute uVisor from the secure memory alias from the
very beginning.