Makefile: Automate source file listings
Source files are found automatically given the list of source folders.
Include paths are listed in a similar way.

In addition, this commit also introduces filters that select source code
based on the MPU architecture. In this way we do not have to maintain
lists of source files depending on the MPU architecture.