Expose all uVisor API via a function table
Instead of calling SVCs in the application directly, the API function in
the uvisor_api table is called, which jumps into uVisors binary and
performs the SVC. This way no synchronization of SVC IDs between uVisor
and the application needs to be performed.

Note that this also speeds up calling the pool queue APIs, since the
position of the export table needs not be manually computed anymore.