Add support for TCMs used as SRAM
In the past a tightly-coupled memory (TCM) could be used easily in place
of the SRAM because what uVisor expected in SRAM was simply moved to the
TCM itself.

Now that we have the page heap the situation is different: If uVisor
uses the TCM then only uVisor own memories and the compile-time-defined
boxes private memories are put in the TCM, whereas the page heap stays
in the actual SRAM.

This commit updates the uVisor codebase as follows:

* Generalize range-checking functions for physical/public SRAM.
* Make sure that sanity checks apply to the correct memory.
* Adapt the static MPU region settings for the SRAM to only cover the
  public SRAM.

This commit requires an update to our linker scripts, where the
public SRAM start and end addresses are specified. The changes required
will be introduced in a different PR and will be captured in the
updated porting guide.