Fix out-by-one error in vmpu_mpu_push
See issue #369.

This function pushes a slot/region into the v7M MPU. If g_mpu_slot is equal to
ARMv7M_MPU_REGIONS_MAX (say 8) then the overflow check

  (++g_mpu_slot > ARMv7M_MPU_REGIONS_MAX)
    ==
  (8 > 8)
    ==
  false

does not fire (when it should) and the function will access the array
g_mpu_priority[8] out-of-bounds and write to an unsupported region number.