Deubg box: Remove outdated comments on reboot
A handler registered to the debug box does not cause the system to
reboot any more. By default, the system is halted, instead.

A user can still request that the system reboots from within a debug box
handler, by calling the `NVIC_SystemReset()` API. This API can only be
called by the debug box when uVisor is enabled.