Box config: Enforce explicit cast for uVisor ctx
The uVisor context pointer is maintained by uVisor at runtime. Each box
finds its own pointer when the box is switched. A nice UX feature was
that the `uvisor_ctx` pointer would magically be a pointer to the custom
type that the user specified when creating the secure box.

With the latest refactoring this feature is not available anymore, but
the user can still define `uvisor_ctx` manually:

```C

struct my_context {
    uint32_t data;
};

\#define uvisor_ctx ((my_context *) __uvisor_ctx)
```

The `__uvisor_ctx` is a `uint32_t` value by default.

This change is not backwards compatible.

Even if it worsens the UX, it makes the context de-referencing more
explicit. In addition, now the same pointer can be used in other files
where a box is active but where the box had not been configured
directly.