vMPU (ARMv7-M): Optimize box enumeration
Previously boxes were ordered based on the link-time order of the box
configuration pointers.

With this commit, the best box configuration is chosen based on the
order that minimizes the total SRAM usage. All possible orders are
computed with a simple permutation algorithm, and for each of them the
total SRAM consumption is computed. The public box is not permuted, as
it's always meant to be the first box to be enumerated.