vMPU: vmpu_acl_stack -> vmpu_acl_sram
This is more consistent with what the function actually does: create an
ACL for the whole box SRAM sections, including the BSS and the stack.

Note: Different MPU architectures implies different implementations of
the BSS and stack sections. On ARMv7-M, the 2 are merged in a single MPU
region, with the help of subregions. On other architectures we use
multiple MPU regions.