main: Remove stale comment about uVisor disabled
This comment can be removed, as it doesn't apply anymore.

In the past, uVisor was disabled by default, while now it is unsupported
by default. This means that this specific code is only executed if the
developer explicitly requires uVisor to be present and linked. Currently
support for what we call disabled mode is broken anyway.