pool_queue: Duplicate pool queue
It's important to run pool queue code on both the S and NS side to
minimize NS to S transitions, and to perform pool queue operations with
NS permissions. Duplicating is obviously not the ideal solution, as
binary size is bloated. Ideally, we could run the same code on both the
S and NS side.