Makefile: Use filters for API sources and core inc
API source files (except for the RTX-specific ones) are compiled by us,
so we can apply the same filters we use in core.

Include directories were not filtered in the past because there was no
need to. With the v8-M target we now have some header files to filter.