vmpu: armv8m: Update slot priority correctly
Previous, after not finding any lower priority slots to evict, we'd
evict the "start slot + 1" slot. But, we'd set the desired priority of
the new slot at the "start slot" slot. Fix this by making and using a
new function that takes a slot parameter, so that we'll always update
the same slot as we program the new region into.