vMPU: Use vIRQ ACL API directly
The level of indirection is error prone, because API signatures can be
changed without raising a compilation error.

This commit simplifies the API so that the IRQ handler is not needed at
all. All IRQs are initialized to the user VTOR. On ARMv7-M, this is done
by us in `virq_init`. In ARMv8-M this is done implicitly since the NVIC
APIs are not wrapped.