ARMv7-M: Fix BSS size count bug
The BSS size of every box was being underestimated. Instead of
accumulating the sizes from the various contributions, we were only
taking the last one.

This problem didn't show up before because the last BSS section used to
be the largest, so the difference would probably go away after rounding.
After changes to the BSS sections, the problem is now more visible.