debug: Disable interrupts before handling debug box
On some platforms, like STM32F429ZIT, there is a HAL tick thing that
runs a lot. If it runs when we are trying to enter the debug box, we
might tail chain to running the tick handler instead of entering the
debug box properly. To fix this, disable all lower-than-SVC-priority
interrupts before entering the debug box. This has the downside that
interrupts can't be used from the debug box while running the debug
handler.