ARMv8-M: Fix scheduler when there is only 1 box
In the previous implementation the context switch function takes the
next box ID before applying modular arithmetic to it. This means that
the very first time a context switch is required, we take the bare
next box ID. Since the next box ID is initialized to 1, if there is only
one box the context switch will fail with an out-of-range box ID error.

This commit makes sure that we modularize the next box ID right from the
first use.