Enable debug print control
When working with a debug version without semihosting the uvisor will
create a Hard Fault, this is a result of trying to send debug prints.
The new API function will set a variable to indicate whether we're
running with or without semihosting.
The source for this problem was discovered in K64F, where there is a
problem probing the debug bit, it is set with or without semihosting.
This fix will automatically apply to all platforms.