pool_queue: Add static assert for management_array offset
In order for the GNU extension of static initialisation of flexible
array member to work. Care must be taken to check whether the offset of
management_array is the size of the structure, so that accessing the
management_array will be to the allocated static memory following the
pool struct. If this assertion fails, accessing management_array might
end of accessing the struct padding inserted by the compiler.