pool_queue: Remove extra unlock in find_first
This unlock in extraneous as the lock will be released at Line 403/416.
This extra unlock will cause synchronisation issues.
Consider 3 threads trying to operate on the same queue.
Thread 0 obtains the lock to and is running find_first.
Thread 1 and Thread 2 is spinning on the lock.
When Thread 0 has found the entry, there is an unlock action. This
allows Thread 1 to operate on the queue. However, the unlock at Line
403/416 executed by Thread 0 will release the lock again. This will
allow Thread 2 to operate on the queue, causing synchronisation problem

Fixes d5c61322c6ba ("pool_queue: Add uvisor_pool_queue_try_find_first")