Add simplified uVisor introduction
This document was designed according to the mbed Handbook specs.
https://github.com/ARMmbed/Handbook/pull/244

The Handbook provides concise descriptions of components in the mbed
ecosystem. However, the uVisor quickstart is too long to fit this
requirement (~400 lines). This document condenses the quickstart into
approximately 40 lines making it easier to understand uVisor at a high
level.